requirements.txt
alembic==1.3.2
Babel==2.8.0
beautifulsoup4==4.8.2
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dominate==2.4.0
elasticsearch==7.1.0
Flask==1.1.1
Flask-Babel==0.12.2
flask-blocks==1.0.20
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language==0.2
guess-language-spirit==0.5.3
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jwt==0.6.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
micawber==0.5.1
peewee==3.13.1
psycopg2==2.8.4
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
redis
requests==2.22.0
six==1.13.0
rq
soupsieve==1.9.5
SQLAlchemy==1.3.12
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
